pyproject.toml
[tool.poetry]
name = "PlayCheckout"
version = "0.1.0"
description = "Play Checkout Challenge"
authors = ["Rodrigo Vaccari <rvaccari@gmail.com>"]

[tool.mypy]
strict = true

[tool.black]
line-length = 120
target-version = ['py37']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | .coverage*
  | manage.py
)
'''

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.4"
django-ninja = "^0.13.2"
django-extensions = "^3.1.3"
python-decouple = "^3.4"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pytest-black = "^0.3.12"
pytest-django = "^4.4.0"
pytest-mypy = "^0.8.1"
pytest-xdist = "^2.3.0"
pytest-cov = "^2.12.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"